Reimplemented naive xmul using constant-time masking
The main reason is so that p::mul can be used more-or-less safely
when best-effort constant-time operations are needed. On the plus side,
since we're already doing a number of shifts, the masking seems to come
mostly for free.

before: xmul/naive_xmul: 6.8442 ns
after:  xmul/naive_xmul: 6.9659 ns

Looking at the dissasembly is a bit intimidating, at least on x86_64:

1. The compiler seems to have completely unrolled the entire loop, at
   least for p64, _and_ vectorized the operations to operate on 4
   shifts simultaneously.

2. The compiler seems to have detected the mask pattern and compiled
   to this repeated operation:

   vpcmpgtq %ymm0,%ymm2,%ymm0
   vpsllvq 0x1d4968(%rip),%ymm1,%ymm1
   vpand  %ymm1,%ymm0,%ymm0
   vpxor  %ymm0,%ymm4,%ymm0

   Fortunately it looks like this interpretation still probably runs
   in constant-time