Tweaked nightly-features feature to not require #![feature] in dependent crates
This was done by moving the low-level xmul definitions into a new,
internal module, so that only this crate needs the crate-level #![feature(stdsimd)]
attribute.

I did explore using external assembly (the best option available on
current stable, rustc v1.55.0). But the overhead of register shuffling
due to the uninlinable function call was too high to be useful.