Added benchmarks over more gf-types, tweaked benchmarks
One place where xmul instructions really shine is in speeding up
gf-types where a table is untenable. So 16-bit gf2p16 up to gf2p64
are interesting to benchmark, though trying to create a table causes
rustc to error with "exceeded interpreter step limit", a hard limit
on const-expr evaluation on current stable.

Another surprisingly interesting benchmark is gf16, which highlights
the costs of the extra checks/masking necessary to enforce that the
type remains a 4-bit value at runtime.