Sped up and improved sanitation of non-register sized (256^2^n) gf-types
The speed up here was adopting slice::get_unchecked for table lookups,
which does result in significant speed up, especially for gf16.
Unfortunately this does add more unsafes, since the finite-field
arithmetic is difficult for the compiler to reason about, but does
end up with indices in-bounds for table lookups.

However this strictly requires the arguments to gf16 and other
non-register sized gf-types be checked on instantiation, or risk
undefined behavior. Fortunately we can generally enforce runtime
checks in Rust's type system.