Minor (2x) optimization to find-irreducibles by skipping multiples of 2
I was hoping to find a more general optimization here, but finding
primes is tricky and not really the main goal of this project.

But, at least skipping multiples of 2 is a no-brainer. It seems like a
minor optimization but doubles the performance thanks to just how many
multiples of 2 there are.

find_p/find_irreducibles_9:   1.0909 us  -36.140%
find_p/find_irreducibles_17:  17.167 us  -45.969%
find_p/find_irreducibles_33:  5.1801 ms  -51.785%