Ripped out io-interworking from RAID implementation
In theory, this was a nice gimmick that would let you easily add RAID to
existing io classes (Read/Write/Seek). In practice it turned into a
lifetime-mess due to needing to share the parity blocks among multiple
readers/writers. There is probably some way to work around this with
increasingly arcane and enigmatic Rust classes, but this is too much
work for a minor feature in a class that's only a proof-of-concept
for this crate!

The last road-block that I stopped at was the split between Arc+RwLock,
Rc+RefCell, and &RefCell. These all solve the specific problem in this
case of needing shared mutation of borrowed storage, but are used in
different contexts depending on thread-safety and the storage's
lifetime. As far as I can tell there's no way to define an API that
supports any form of shared-mutable-storage without creating your
own trait and implementing it for all possible shared-mutable-storage
classes.

But anyways, that's gone now, and in it's place is just the root
algorithm of RAID's single and double parity. This form is much better
anyways since it is completely no-std compatible, and, while requiring
more manual slice manipulation, is a bit more flexible.