Added p/gf-types slice_from_slice and slice_from_slice_mut
Newtypes in Rust are unfortunately quite annoying to deal with when in
slices, however since these types are repr(transparent), it's safe
and basically free to transmute their slices. After rewriting this
simple transmute wrapper in enough places, I figured it should be a
first-class citizen of the repr(transparent) types.

Though I wish Rust had some mechanism to enable this for free on
repr(transparent) types, it seems that converting slices is a
common pattern for these types and there's not way to do so
efficiently without a bit of unsafety...