Reworked LFSRs after some misunderstanding with how they map to Galois-fields
Given an LFSR:

    <-.- 1 <-- 1 <-- 0 <-- 0 <x- 1 <x- 1 <x- 0 <-- 0 <-.
      '-----------------------'-----'-----'------------'

I initially thought you could use the entire state of the shift register
as output for pseudorandomness generation. This isn't true and leads to
noticable patterns, especially when there are few taps (which happen to
be the cheapest to implement in hardware).

While researching more into what would be the best LFSR polynomials to
provide by default, I learned that the best way to get randomness out of
an LFSR is to shift the state for every bit of output.

Both of the methods have the same cycle length of all non-zero integers,
which is a bit interesting.

---

Fortunately, it turns out this maps directly to polynomial division, in the
same way the state update maps to the polynomial remainder.

This lets us generate whole words at a time, so performance isn't
significantly hurt, though Barret reduction is less effective since, as
far as I know, there isn't a way to implement division with the same
technique.

---

Unrelated, but something else I realized while reading up more on LFSRs
is that they are cheap (O(loglog n) to skip around in, and can even be
ran backwards fairly easily. So I've added those features to the
implementations.

I've also removed the Rng trait from the example, but will readd it into
the generated Lfsr structs. It's useful, but not that interesting for
the example, as the randomness seems to take a backseat to highly
manipulatable state positions.