LFSR refinement: Now with reflections and odd-sizes
This was actually quite a bit less head-scratching than odd-sized next
calls took get working, probably mostly thanks to having figured out
most of the details for odd-sized CRCs. It mostly just took a
significant of iteration.

Also reflection here is just implemented as an optional last-stage
bit-reverse. This probably isn't optimal, but it was easier, less
error-prone, and this really is a small part of the whole library...