Added support for partial CRCs
So now CRCs can be resumed, which is actually the only reason for the
result xor as far as I can tell. I've also deduplicated the init and xor
options, since the init argument to the crc function can be used to
differentiate the two, though it's incredibly rare as far as I can tell
for these to not be equivalent.