Fixed most inter-doc links
Normally, all of these links would be relative links that can be
resolved locally when building local docs. But unfortunately, with half
the documentation living on docs.rs, and half living on GitHub, this
simply isn't possible. This commit resolves these intra-doc links with
absolute URLs, which is fragile and will stop resolving correctly after
new versions (these links uses "latest" and "master"), but unfortunately
I'm not aware of a solution to this.

Relative links are used where possible, though it's also interesting to
note that Rust's "item" links have a difficult time resolving when the
target is generated by a proc_macro. I can't say this is too surprising,
since proc_macros generally break Rust more often than not.

There's a lot of links unfortunately, so I wouldn not be surprised if I
missed one or two. At least `make docs` no longer prints a bunch of link
warnings.